Refactor Pets page, display given pet, move create Pet form to separate view, add toggl button to navigate between view pets and add pet pages
